requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.3.0
Django==3.0
djangorestframework==3.11.0
importlib-metadata==1.3.0
gunicorn==20.0.4
kombu==4.6.7
more-itertools==8.0.2
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
supervisor-logging-gelf==0.0.1
vine==1.3.0
zipp==0.6.0
